Revert to full dataset (#381)
* remove old make cmds

* uv in parse data script

* change default zone id dtype

* update to full dataset binaries

* Refactor cleanup method in FileCoordAccessor to safely close resources

* revert to original test cases

* Enhance cleanup methods in AbstractTimezoneFinder

* bugfix check_speed_timezone_finding script

* Update default input path in configs.py to use combined-with-oceans.json

* update documentation

* add .cursorrules

* Remove destination path argument from file_converter.py call in parse_data.sh

* version 8.2.0

* add Python 3.14 test support

* Update CI configuration to support PyO3-based wheels for Python 3.14

* Update CI workflow to enable PyO3 ABI3 compatibility for building wheels on Python 3.14

* Update .cursorrules to clarify command prefix usage in documentation

* improve integration test performance

* introduce data dependency group

* update the lock

* Update documentation to specify lockfile management for dependency changes

* improve global function test performance

* Add shared fixtures for pytest to improve test organization and performance

* Refine documentation in .cursorrules to emphasize the importance of pre-commit hooks in the code changes workflow

* remove PYO3_USE_ABI3_FORWARD_COMPATIBILITY

+++ timezonefinder/configs.py
@@ -85,5 +85,5 @@ def available_zone_id_dtype_names() -> Tuple[str, ...]:
     return tuple(sorted(_ZONE_ID_DTYPE_ALIASES))
 
 
-DEFAULT_ZONE_ID_DTYPE_NAME = os.getenv("TIMEZONEFINDER_ZONE_ID_DTYPE", "uint8")
+DEFAULT_ZONE_ID_DTYPE_NAME = os.getenv("TIMEZONEFINDER_ZONE_ID_DTYPE", "uint16")
 DEFAULT_ZONE_ID_DTYPE = get_zone_id_dtype(DEFAULT_ZONE_ID_DTYPE_NAME)
